move sysdep files to new location
