+++ haslibutilh.c
@@ -1,0 +1,6 @@
+/* Public domain. */
+#include <libutil.h>
+
+int main(void) {
+    return 0;
+}
