+++ haslimits.c
@@ -1,0 +1,6 @@
+/* Public domain. */
+#include <limits.h>
+
+int main(void) {
+    return 0;
+}
