+++ hasutilh.c
@@ -1,0 +1,6 @@
+/* Public domain. */
+#include <util.h>
+
+int main(void) {
+    return 0;
+}
