+++ hasutmpxsyslen.c
@@ -1,0 +1,8 @@
+/* Public domain. */
+#include <utmpx.h>
+
+int main(void) {
+
+    struct utmpx ut = {0};
+    return ut.ut_syslen;
+}
