+++ old/sysdep/haslibutilh.h-yes.c
@@ -1,6 +1,1 @@
-/* Public domain. */
-#include <libutil.h>
-
-int main(void) {
-    return 0;
-}
+../../haslibutilh.c
