+++ old/sysdep/haslimits.h-yes.c
@@ -1,6 +1,1 @@
-/* Public domain. */
-#include <limits.h>
-
-int main(void) {
-    return 0;
-}
+../../haslimits.c
