+++ old/sysdep/hasmlock.h-yes.c
@@ -1,9 +1,1 @@
-/* Public domain. */
-#include <sys/mman.h>
-
-int main(void) {
-    char x[16];
-    mlock(x, sizeof x);
-    munlock(x, sizeof x);
-    return 0;
-}
+../../hasmlock.c
