+++ old/sysdep/hasutilh.h-yes.c
@@ -1,6 +1,1 @@
-/* Public domain. */
-#include <util.h>
-
-int main(void) {
-    return 0;
-}
+../../hasutilh.c
