+++ old/sysdep/hasutmp.h-yes.c
@@ -1,10 +1,1 @@
-/* Public domain. */
-#include <time.h>
-#include <sys/time.h>
-#include <utmp.h>
-
-int main(void) {
-
-    struct utmp ut = {0};
-    return 0;
-}
+../../hasutmp.c
