+++ old/sysdep/hasutmpname.h-yes.c
@@ -1,11 +1,1 @@
-/* Public domain. */
-#include <time.h>
-#include <sys/time.h>
-#include <utmp.h>
-
-int main(void) {
-
-    struct utmp ut;
-    if (sizeof ut.ut_name < 8) return 111;
-    return 0;
-}
+../../hasutmpname.c
