+++ old/sysdep/hasutmpxsyslen.h-yes.c
@@ -1,8 +1,1 @@
-/* Public domain. */
-#include <utmpx.h>
-
-int main(void) {
-
-    struct utmpx ut = {0};
-    return ut.ut_syslen;
-}
+../../hasutmpxsyslen.c
