+++ old/sysdep/hasutmpxupdwtmpx.h-yes.c
@@ -1,13 +1,1 @@
-/* Public domain. */
-#include <utmpx.h>
-
-static void foo(void) {
-    struct utmpx ut = {0};
-    updwtmpx("/nonexistent", &ut);
-}
-
-int main(int argc, char **argv) {
-    if (!argc)
-        foo();
-    return 0;
-}
+../../hasutmpxupdwtmpx.c
